Activity Log: Continue fetching multi-page results (#22831)
* Activity Log: Continue fetching multi-page results

It has been possible that we have more results in a request to
the Activity Log than would fit in the requests number of events
for a single page. Previously we have been ignoring everything
past the first page.

In this patch we're looking at the API results and if there are
more pages of results we will start a loop to fetch those in
sequence by passing the `page` parameter for what should be the
next page.

* Rebuild based off of `nextAfter` being returned by API

After reworking the backend to remove the complexity brought in by using
both paging and `search_after` in the queries this patch removes the use
of paging in Calypso and simplifies the keys for polling.

Previously we were requesting follow-up pages but in this revision we're
continuing to request based solely off of the presence of `nextAfter` in
the response.

Action meta disambiguates polling vs. next-page-fetching.

Also we have to do something fishy in order to make polling reliably do
what we want. We alternate between fetching 1 and fetching up to 100
events on the polls in order to force the presence of `nextAfter`. That
is, if we poll for new events and request 100 but only one comes in then
there will be no `nextAfter` value so we trim it to a limit of 1 and
then that value _will_ be present. Once we get an event we open up the
limit again to 100. If no new events appear then it collapses back to 1
because no events came over and we'll reuse the previous `nextAfter`
until a new event appears.

* Add missing explanatory comment