Framework: Break chunk dependency cycles
This renames section-preloader to section-info and adds the ability to get a section by name. This works by firing a filter hook that the sections-middleware processes and returns. This may seem a bit obtuse, but it lets us break a chunk dependency cycle.

Before this patch, to get access to section settings, code was importing sections-middleware, which imports client/sections, which contains all of the dynamic imports for our sections. This would cause the chunk that needed the section settings to become a parent to all of our sections, which created a cycle in the chunk dependency grath.

You could see this in the output of npm start. Every section chunk depended on build, which is expected, but also plugins, wp-job-manager, and zoninator.

This cycle slows down chunk analysis and makes optimization more difficult.