JP Settings (data-layer based): Keep Track of Module Activation State (#23855)
This part of state isn't used by selectors (yet). However, the reducer does listen to those action types.

Reducer and tests basically duplicated from `state/jetpack/settings`, with some minor modifications (to work with `keyedReducer` and Jest). This is the same strategy as with #23774 -- duplicate temporarily so stuff is readily available when switching over.