Activity Log: Sort events using `rewindId`
Previously we have been sorting Activity Log events by the `activityTs`
value which comes from the `published` attribute in the events.

Problematic is that this timestamp is a truncated version of the
timestamp and when events happen in short succession then we can get
events out of order.

By changing to the `rewindId` we move to a canonical value that should
preserve ordering to the best ability of the remote system where the
value was first generated.