Center Spinner position between the form and CTA button.
The bottom distance of the Spinner to the CTA button was shorter than the top one to the form field.
This change removes bottom margin of the password field and sets the distance using CTA button margin.
Moving the Spinner to a div containing the button, allows to moderate the bottom distance to the Spinner.