Plugin Notices: simplify the removePluginsNotices function
Instead of passing it the array of notices to remove, pass names of the status types,
i.e, 'completed', 'inProgress' or 'error'. We always remove all notices of the same
status anyway.

The benefit is that we can stop passing a `notices` prop around at many places, cleaning
up the code.