Stop polling for "new" Activity Log events
For now this doesn't make much sense as we start to talk about filtering
and searching. We need to think of a more semantically-matched way to
grab updates. Eventually a server-push model fits the paradigm best here
but that work will have to take place in further PRs when the project is
a bit further along in the process.