Activity Log: Show existing Rewind operations (#24079)
Calypso is using two systems to gather information about running
Rewind operations. Once it kicks off a Rewind session it starts
hitting the `restore-status` API endpoint and displays info about
that endpoint in a banner.

We're also relying on the Rewind "State machine" endpoint to gather
information about the entire Rewind system. This API endpoint
already provides information on active Rewind sessions.

One problem we've had is that Rewind operations don't show their
banners after page refreshes and in other sessions. If you visit
a page on a site where a Rewind is already going on it will look
like you can't start a Rewind operation and there will be no
explanation.

In this patch we're using the response from the Rewind API to kick
off a request to the `restore-status` API so that if there is a
Rewind operation going on we can see it immediately.