JPC: provide extra section to Jetpack plans and features in plans pages (#24077)
* JPC: add section to plans landing page with plan links to jetpack.com info.

* JPC: add tracking to the buttons redirecting to external information.

* JPC: include the extended plan info section in the plans page at `/jetpack/connect/plans/:connected_site_slug`.

This will show the additional section for freshly connected sites.

* JPC: create new stateless component to inject extra info section to plans pages.

This removes code repetition in defining the extra component both in plans-landing
page for unconected sites, ands plans.jsx for connected ones.

* JP: include updated image for Jetpack start.

Will be updated in https://github.com/Automattic/wp-calypso/pull/22957,
this commit adds it for proper testing.

* JPC: correct typo in header description of the new section.

* adjust styling

* Updated some styling and made CTA buttons sentence case.

* Remove redundancy in styling and address lack of responsiveness.

* remove redundant css attributes
* make redirects to open in a new tab
* rename onClick to alternative since it does not pass event object
* remove redundant className
* update illustration.

* JPC: include padding to the illustration placement.

* update plan test snapshot

* JPC: adapt styling of the extra section.

Plans grid is not fully responsive and it is not possible to match
side margins well for all views. In order to test this addition,
we modify the size of the card with extra sections and it content,
to be narrower and behave "well" for all viewports.

* JPC: remove dynamic creation of track events from extended info section.

This is done to efficiently search for them in the codebase.

* JPC: polish the styling of the external info section for plans.

* reduce top margin
* increase the container size

* Correct tracking prop name when calling the PlansExtendedInfo component.

The prop name onClick was incorrect and did not correspond to the component`s prop name.

* Unify tracking function definitions between the two JPC plans views.

Correct the tracking parameter definition, which differed between the two. Also, adjust the definition to return the function that records event,
instead of recording the event on render, which results in event firing on render of the component it is passed to.

* Update test snapshot.

* Simplify function syntax and add props definition to the component.