Enable video thumbnail editing on private sites.
A bug was discovered that prevented image editing from warking on
private sites. As a quick fix the edit button was removed from the media
diolog when the site was private. This also removed the edit button for
wideos, which aren't affected by the same issue.

This change checks the mime type of the item being edited and only hides
the button for images on private sites.