Activity Log: Discard `isDiscarded` (#24016)
We have been conditionally rendering events in the Activity Log based
on whether a previous Rewind operation (restore) has _undone_ the
appearance that the event ever occured. That is, imagine that we restore
a site to a point before a post was published. For all practical
purposes that post was never actually published; it would appear in a
disabled state as "discarded" by the restore.

This UI feature is helpful but as we start to implement the ability to
perform partial restores on a site the complexity involved in
maintaining this property surges. Events depend on specific parts of
a partial restore and it's not trivial to know what parts those are.
That is, the post is not discarded by a theme restore, but it would be
discarded by a database restore. Likewise a media item depends on both
the database _and_ the filesystem restores.

Therefore we have decided to remove this calculation for now and stop
showing it in the UI. At some point in time we may bring it back but for
now for practical purposes we are eliminating it in order to remove the
obstacle for moving forward as a project.