Activity Log: Linearize and page view based on perishable filters (#23898)
Activity Log: Linearize and page view based on perishable filters

This is a dramatic change to the Activity Log display - it removes the monthly/daily list view and linearizes it instead. It removes complicated persistence going on to maintain internal state when updating the events query. It uses paging of already-requested values instead of infinite scrolling with live fetches.