Post Editor: rename and split mergeIgnoringArrays into two functions
Rename the `mergeIgnoringArrays` to a more appropriate name: `mergePostEdits`.
It's a function that merges two post edits into one.

Also create a function called `applyPostEdits` that applies a post edits object
to a post objects and returns a modified post.

At this moment, both implementations are identical, but that will change when
editing post metadata is intruduced into the Post Editor Redux store.