Activity Log: Honor site timezone settings in date header
Since the introduction of the linearized view we have been showing
date headers at the top of the list of activities and whenver a
subsequent event is on a new date. We have two issues with this
first iteration:

 - We no longer show "today" when today is the date
 - We don't respect the timezone of the site

In this patch we're fixing these issues and adjusting all of the
timestamp values based on the site's offset value.