Refactor the rendering of importer components in the `SectionImport` component
Previously, adding a new importer required adding a fair bit of boilerplate in
the render method of the `SectionImport` component. This commit refactors things
so that to add a new importer the developer simplay adds it to the `importers`
array.

There are now some new new methods in SectionImport:
- `renderImporters` replaces the previous `getImports` method. The new method
returns an array of react elements representing the importers that are rendered.
It calls either `renderIdleImporters` or  `renderActiveImporters` depending on
whether the store contains any import data.
- `renderIdleImporters` returns an array containing each enabled importer at a
given state. It is currently used to render the importers while the page is
loading and also while there are no active import jobs.
- `renderActiveImporters` receives the list of in progress import jobs and
renders just the importers for those jobs (as long as the importer is enabled).

Before any import data is passed to the `renderActiveImporters` function it must
be filtered using the `filterImportsForSite`, to ensure the import data is only
for the site being viewed.