Components: use flexbox for rendering content of `PurchaseDetail` component. (#24262)
* Components: use flexbox for rendering content of the PurchaseDetail component.

This change is done to fully control the layout of the component with minimal tweaking.
Also, to make it fully responsive.

* PurchaseDetail - updated flexbox styles to match production visually

* Components: apply custom icon styling to gridicons in PurchaseDetail component.

There are two types of icons displayed on the feature cards.
Previous commit applied styling to custom icons, which contain
illustrations. Here we mimic the styling for gridicons.

* PurchaseDetail - break long strings in the button to prevent layout breakage

* PurchaseDetail—extend the text container to keep he icon at the end of the parent flex component.

* ProductPurchaseFeaturesList: temporary fix to the overflowing border of the left hand side card.

This is an edge case occurring in Chrome and is a known issue (not due to changes in this patch).