Infinite List: run the ES6 class codemod
The InfiniteList class was skipped during the React 16 migration because it assigned
a class method in a nontrivial way:
```js
smartSetState: smartSetState
```
After a manual intervention and running the codemod again, the component is now ES6.

I also added a few comments to mark instance methods that are caused from outside, by
a parent component via a ref.