Ad Tracking: Facebook Advanced Matching - email, first and last name
This commit adds the e-mail, first name and last name to the data that
is being sent to Facebook, such that ads can be better targeted.
The hashing happens before existing the user's browser
(handled internally by the FB library).

Since first name and last name are only available with settings,
it needs to wait for a response from this endpoint before initializing
the FB tracking pixel.