Post Edit Store: move metadata update logic from action creators to the store (try 2) (#24376)
* Post Edit Store: move metadata update logic from action creators to the store

Currently, the `handleMetadataOperation` function (which plays a role similar to a
Redux action creator) retrieves the metadata of the edited post, applies all edits
to them and dispatches a `EDIT_POST` Flux action with the updated metadata list.

This PR shuffles the logic around a bit. The `EDIT_POST` action doesn't contain
the full updated `metadata` list, but only the list of edits. Then the `set` function
in the `PostEditStore` applies these edits to the post.

The end result is the same, but organizing the code this way will help a lot with
Redux refactoring, as it will allow the Flux and Redux implementations of metadata
editing to live together for a while.

Tests were updated: some functionality that was previously tested in `PostActions` is
now present in `PostEditStore` tests.

* Post Edit Store: guard for null attributes arg when setting edits on a post

The `EDIT_POST` Flux action can be dispatched with null `post` property. Don't
throw a `TypeError` in that case.