Plans: Use PlansSkipButton for WP.com Plans (#24333)
Previously (as of #22384), we were using the entire `plans` array as passed to the `PlanFeatures` component, and simply filtering to yield only the free plan. We would then pass a number of attributes obtained that way to the `PlanActions` component, which would render the 'Start with Free' button. Out of all the props passed to that component, the only relevant one seems to be `onUpgradeClick` -- the click handler (which is somewhat of a misnomer for free plans).

Still inside the `PlanFeatures` component, we would then move on and create a React portal to place the button at the bottom of its parent `PlansFeaturesMain` component (this is what the `PlanFeaturesBottom` component was for).

This PR uses the `PlansSkipButton` component instead, rendering it directly into `PlansFeaturesMain`. However, since we're no longer going through the `PlanFeatures` component, we lose the [additional wrapping](https://github.com/Automattic/wp-calypso/blob/34755834a2d4b43b58b0f6cf572bde8f886449a6/client/my-sites/plan-features/index.jsx#L819-L831) done to the `onUpgradeClick` prop in its `connect()` statement -- which essentially just binds the selected plan as a cart item to it. However, we can easily compensate for that by passing `null` to `onUpgradeClick`, which is the value that represents the free plan.