Purchases: add redirectLoggedOut to all paths
Currently if you visit a route like `me/purchases` when logged-out, you
will see a blank page (see #23785 for the general problem).

This will allow all paths in the `me/purchases` routes to redirect to
the login page if logged-out, and then redirect to the page originally
requested.

In other words, with this PR, visiting `/me/purchases` will instead
redirect to the login page and, once logged in, to the purchases page.