Post Editor: reduxify editing the post excerpt  (#24326)
* Post Editor: reduxify editing the post excerpt

Extract the excerpt editor into a separate component (it was inlined into
`EditorDrawer` until now) and Reduxify it on the way: get the post data
from Redux and dispatch Redux action on edit.

* Post Editor: when restoring a revision, copy post excerpt and title to Redux

When restoring to a specific post revision from history, we rewrite the current
post's title, excerpt and content. The title and excerpt properties are edited
in Redux, so the update should be performed there. The only property that remains
Flux is `content`.

* Post Editor: don't translate excerpt placeholder twice