Activity Log: Add styling to threat alerts
This feature currently sites behind the `rewind-alerts` feature
flag in `development` and is part of a bigger project to add
threat alerts to the Activity Log.

In this patch we're adding some styling and class names to the
existing HTML, rearranging some HTML, and introducing the helper
component `<DetectionTime />` to unify how we display the time
since a threat was first detected.

These are mainly style-only changes on the existing "shell" of
the threat system so no changes should be seen by any customers.