Remove unused Post Editor code (#24429)
* Post Editor Visibility: remove the showingAcceptDialog instance property

The property is being set, but never read from. Looking at the initial OSS commit, the
`EditorVisibility` UI used to be a modal popover (it's a popup menu today) and the property
was used to cancel hiding the popover when an accept modal dialog is shown. Today, the UI
is different and this property is not needed.

* Editor Action Bar: remove the unused onPrivatePublish prop

The Action Bar doesn't do any publishing work and doesn't use the prop.

* Post Editor: remove the legacy post-editor/revisions config options

It's been enabled in all environments for a few months. We can remove some
dead code, like the `EditorRevisionsLegacyLink` component that was showing
a link to wp-admin.

* Post Editor: move isPrivateSite prop to EditorVisibility's mapStateToProps

The `EditorVisibility` component can get the `isPrivateSite` flag in its own
`mapStateToProps` function. It doesn't need to be the responsibility of the
parent component to pass it. Refactor accordingly.

* EditPostStatus component doesn't need a 'site' prop any more

It was used by the `EditorRevisionsLegacyLink` component which was removed,
and to get the `is_private` option which was replaced by the `isSitePrivate`
selector.

* Remove unused code from EditorVisibility component

The `onPasswordChange` method is not used and can be removed, together with the `react-dom`
import and a `postPassword` ref.

Also, the `onKey` method no longer exists.