Add unified diff patch viewer (#24508)
This patch creates a component to visualize a unified diff output (created with commands such as git diff and diff -u). It shows line-level changes to the compared contents. It does not do the comparison but instead relies on jsdiff to parse an existing patch file. It's plausible that in the future we expand this component to alternatively do the comparison but for now this behavior is not needed.

Further, it's plausible that we extend this to mark changes from within the component as well. That's a more complicated task than showing changed lines so will not be scoped into this PR.