Use Redux to store copied metadata when starting to edit a post copy
When starting to edit a post copy, we copy some attributes to the Flux store
and some to the Redux store, depending on how exactly editing of that particular
attribute is implemented at the moment.

This patch moves the copied metadata to Redux. The only metadata we copy right
now are the geolocation fields, and these are being moved to Redux in the next patch.