Post Editor: reduxify editing post sharing options
All components in the `editor-sharing` folder are now self-sufficient when it comes to
getting the `site` and `post` object. Instead of passing them down as props, they get
them from Redux in their `connect` calls.

The only change that is not 100% straightforward is the `PublicizeMessage`. Here, there is
additional change that the component itself doesn't dispatch any metadata update actions, but
always relies on the `onChange` handler passed from above: the action dispatch is moved to the
parent.