Post Editor: save post status updates to Redux when changing post visibility
When the `EditorVisibility` component updates the post status, make sure that the
`status` attribute edits are stored both to Flux (as they always were) and to Redux
(not stored until now). That ensures that in further refactorings it doesn't matter
which store we read `status` from: it's always in sync. That makes a lot of things
easier.

The `updatePostStatus` method was used only at one place, so we inline it into
`updateVisibility`.