State: Use getHttpData() for geolocation data
In #22549 we added the `getHttpData()` abstraction
for simple network data needs.

In this patch we're moving the `geo` state subtree over to this
new abstraction and removing the old code it previously required.

This approach is build on the data-layer ideas and extends it
to store data outside of Redux.