Reduxify setting the post geolocation (#24563)
* Add Redux logic for editing post metadata

Add two new actions creators, `updatePostMetadata` and `deletePostMetadata` that
produce `POST_EDIT` actions with the appropriate properties. Equivalent to the
`updateMetadata` and `deleteMetadata` Flux actions.

The `state.posts.edits` reducer now correctly merges the incoming post updates
into the metadata edits, leaving only the unapplied ones.

The `isEditedPostDirty` selector compares the outstanding edits to the original
post metadata values and determines if the edits are noops or not, affecting the
return values of the selector.

All new Redux logic is covered by unit tests.

* Use Redux to store copied metadata when starting to edit a post copy

When starting to edit a post copy, we copy some attributes to the Flux store
and some to the Redux store, depending on how exactly editing of that particular
attribute is implemented at the moment.

This patch moves the copied metadata to Redux. The only metadata we copy right
now are the geolocation fields, and these are being moved to Redux in the next patch.

* Reduxify setting the post geolocation

Use the new Redux metadata logic to edit post geolocation: get data from Redux
selectors and do edits with Redux action dispatches.

* Send geolocation updates as formatted strings to ensure correct comparison

Formats float values of geo coordinates to strings with 7 decimal digits before
sending them as metadata values to REST API. This ensures correct comparison between
the values in the edits and the values returned by REST API responses. The responses
use the string format described above.

Without this patch, geolocation updates are never marked as "saved", because the
values are different.

* When copying post metadata, ignore the 'id' attribute

Metadata edits that have an 'id' attribute are ignored by the server and
the edit never gets saved.