PublicizeMessage: store the userHasEditedMessage flag as instance property
Fixes a bug where Redux update of the `message` prop interacts badly with the
`setState( { userHasEditedMessage } )` call. That `setState` leads to two
rerenders of the component: first with new `userHasEditedMessage` value and old
value of `message`, the second with new value of `message`.

There is a following sequence of events:
1. The textarea has value "a"
2. User types "b"
3. The DOM value of the textarea becomes "ab"
4. First rerender happens and sets the value back to "a"
5. Second rerender happens and sets the value to "ab"

This back-and-forth value setting resets the position of cursor.