Checkout: redirect logged-out users to login page
When logged-out, a blank screen is displayed on the all checkout pages but `/checkout/:domain/:product?`.
This trouble is caused by #22993 which fixed a bug filed in #22857 (see #23785 for the general problem).
For example, renewal email was also affected by this bug (see p9jf6J-rI-p2).

This PR fixes an empty page displayed to logged-out users when they try
to access any `/checkout` pages.