Framework: `getHttpData` use dispatch queue until initialized
Before the `getHttpData()` middleware is injected into the Redux
store it cannot dispatch any actions. In the first draft of this
code the actions were overlooked. In this update we are queueing
actions until the dispatcher becomes available then flushing out
all the actions after a small delay after the enhancer loads up.