Post Editor: save post status updates to Redux (#24698)
* Post Editor: save post status updates to Redux when changing post visibility

When the `EditorVisibility` component updates the post status, make sure that the
`status` attribute edits are stored both to Flux (as they always were) and to Redux
(not stored until now). That ensures that in further refactorings it doesn't matter
which store we read `status` from: it's always in sync. That makes a lot of things
easier.

The `updatePostStatus` method was used only at one place, so we inline it into
`updateVisibility`.

* Post Edits Reducer: remove saved status edits from local edits

When receiving a request to change the `status` attribute, the server
treats `publish` and `future` as synonyms. It's really the post's `date`
that determines the resulting status, not the requested value.
Therefore, the `status` edit is considered saved and removed from the
local edits even if the value returned by server is different.

All other statuses (draft, private, pending) are 1:1. The only possible
exception is requesting `publish` and not having rights to publish new
posts. Then the server sets a `pending` status. But we check for this case
in the UI and request `pending` instead of `publish` if the user doesn't
have the rights.