Post Editor: improve names of selectors that check if edited post is password protected
There are two selectors (added in #16403) that check if the edited post is password protected
and if it has valid (non-empty) password set:
- `isEditedPostPrivate`
- `isPrivateEditedPostPasswordValid`

They are used to enable/disable the "Publish" button in the Post Editor UI.

Their names, however, are misleading. A "private" post (visible only to site admins and editors)
is something quite different from a "password protected" post.

This PR changes the selector names to:
- `isEditedPostPasswordProtected`
- `isEditedPostPasswordProtectedWithValidPassword`

There will be soon a `isEditedPostPrivate` selector that checks for the real "private" status.