Post Editor: remove PostEditStore.isNew() usages
At most places, usages of the Flux getter `PostEditStore.isNew()` can be replaced
by the Redux selector `isEditorNewPost` that returns the same value. Namely, this
patch introduces the Redux selector into the following components:

- `EditorAuthor`, where it's used to show author of a new post, which is 100%
certain to be "you", even if the post to edit is still loading. If the post is
not new, the component displays a placeholder while loading.

- `EditorDiscussion`, where it shows the default discussion options, as determined
by site settings, for a new post that doesn't have any explicit discussion options
set yet.

Then we remove `isNew` prop from a few components that just pass it to children:
`EditorActionBar`, `EditorDrawer`, `EditorSidebar`.

Then there are a few usages of `PostEditStore.isNew()` in the root `PostEditor`
lifecycle methods that react to changes in Flux or sync Flux changes to Redux.
Here we still need to retrieve the information from Flux. To achieve this, we
make use of the fact that "post is new" if and only if "there is a saved post
that has an ID set". New posts don't have an ID yet.