Activity Log: Add state for filtering the log
@see p9rlnk-b5-p2

As we start looking to implement filtering and querying in
the Activity Log we need a place to store state related to
which filters and what queries are currently active.

In this patch I'm adding a new reducer and associated
selector for the "filter state" and storing the selected
page number in there. The page number does not affect the
query sent to the Activity Log API but it does determine
which range of the results to show in the view.

I have designed the reducer to function from two basic
actions: set, which takes the default initial values and
overrides them with values from the action; and update,
which operates in the same way but instead of starting
with the defaults it starts with the previous state.

This allows us to _set_ properties from the URL when we
decide to do that but it also allows us to make atomic
changes from things like clicks and text inputs. A reset
thus is performed by calling `set( {} )` which will
return the state to the default.