Post Editor: move Redux action dispatches on post load/save to lib/posts/actions.js
When the Post Editor loads a post to edit (RECEIVE_POST_TO_EDIT) or receives a response
to save request (RECEIVE_POST_BEING_EDITED) then we dispatch a few Redux actions to
keep Redux in sync with Flux.

This patch moves the dispatches to the action "thunks" in `lib/posts/actions` and uses
the Redux bridge to perform them. Previously they were in `PostEditor` lifecycle methods,
which makes the code hard to follow and reason about. Further Redux migration will also
be much easier now.