Ad Tracking: Nanigans - ignore 0 or negative value carts
Before this change, 0-value carts (i.e. when user already had a plan, but claimed a domain) were also considered a conversion in the Nanigans system. We have already decided to remove such carts for the FB tracking, so we are replicating the same behavior to Nanigans.

After this change, negative carts (i.e. carts that may have received back credit) and the ones with 0 value (i.e. domain purchases when the user already has a plan) will not be recorded.