Activity Log: Fetch site alert data
In #24729 we have largely built out the UI for showing
malicious threats on a site but up until now we have
had no way of loading this data into Calypso.

In this patch we're using the `getHttpData()` system to
fetch that data from the new and developing API on
WordPress.com and hooking it into the React component.