Allow toggle of public display of geo data
The wp-admin UI for geo-tagging posts on WordPress.com sets a
geo_public meta field in addition to the geo_latitude and geo_longitude
fields currently set by Calypso.  Without geo_public present, the WP
frontend does not display geo-location information, as reported in
issue #4099.

This commit adds a checkbox to the UI to allow authors to toggle this
public display.  To maintain consistency with prior versions of Calypso,
this value defaults to keep the geo-tagging information private.