ESLint: add ignoreRestSiblings flag to the no-unused-vars rule
Allow unused variables that are used to omit properties from an object:
```js
const { a, b, ...rest } = props;
```
`a` and `b` are unused variables, but yet ESLint doesn't complain about them,
because they are used to omit properties from the `props` object and assign
the new object to `rest`. That's a handy alternative to Lodash `omit`.