Activity Log: Show only basename of file in threat alert
The threat alert card header shows a terse description
of the kind of threat that VaultPress has detected.
In this patch we're refining the display so that when
we show a filename we only show the `basename()` of
that file until you open the card and see the entire thing.