Framework: Remove unused v1 Circle CI config
In #23131 we merged in the work that @levlaz created
for us, which was to update our config to the v2
format. This upgrade was entirely motivated at first
by the bug that we worked around in #21518.

Since then our v2.0 config has been working fine and
it's time to remove the old one and the workaround
file since the bug in v1 was fixed.

In this patch we do just that.