EditPostStatus: default post status to 'draft' when passing prop to EditorVisibility
When editing a new post, the Redux state doesn't have the `status` attribute and the
`status` prop passed to `EditorVisibility` is undefined. When updating post visibility,
e.g., from public to password-protected, the component erroneously updates the post
status to `publish`.

The net result is that the post autosave doesn't work: the post has no ID and yet its
status is `publish`. That's an impossible combination and the `autosave` action will
return a `NO_AUTOSAVE` error.