Post Editor: refactor redirect to a route with the correct post type (#24910)
* Post Editor: refactor redirect to a route with the correct post type

After the post editor loads the post to edit, specified by ID in the route params:
```
/page/example.blog/123
```
it checks if the post with ID=123 is actually a Page. If it's not, it redirects to:
```
/post/example.blog/123
```
or, in case of a custom post type, to:
```
/edit/jetpack-testimonial/example.blog/123
```

This PR refactors the code that does this redirect. It's moved from `PostEditor` lifecycle
method to the `startEditingExisting` flow in the controller. This makes the sequence of
events much clearer:
1. Start editing existing post with a gived ID
2. Wait until the post to edit loads from server
3. Check the URL type and actual post type and redirect if needed

* Don't redirect if the browser navigated to another route while load was in progress

* Extract context.path to its own variable