EditPostStatus: Get the post data from Redux and remove unneeded props (try 2) (#24947)
* EditPostStatus: Get the post data from Redux and remove unneeded props

The `EditPostStatus` can get all the post attributes it needs from the
Redux-supplied `post` prop and doesn't need extra props like `type`,
`status` or `isPostPrivate`.

Since #24698 was merged, the post status is stored in Redux and always
has the correct value. There's no need to retrieve the `status` from
the Flux store.

The same applies to the `isPostPrivate` prop, whose value depends on the
status. Instead of an extra prop, determine the `isPrivateOrPasswordProtected`
value from the `post` object. Exactly the same value is used also by
`EditorActionBar` for exactly the same purpose: to determine if the post
supports the `sticky` attribute and if the toggling UI should be displayed.
For `EditorActionBar`, similar Redux refactor was done in #24760.

There are also two props being removed that `EditPostStatus` doesn't use: `site` and
`onTrashingPost`.

This PR touches some JSX that has a preexisting a11y eslint warning about labels
not being associated with their controls. That's for another PR to fix. Don't be
surprised if `eslines` reports this PR as a failure.

* EditPostStatus: default post status to 'draft' when passing prop to EditorVisibility

When editing a new post, the Redux state doesn't have the `status` attribute and the
`status` prop passed to `EditorVisibility` is undefined. When updating post visibility,
e.g., from public to password-protected, the component erroneously updates the post
status to `publish`.

The net result is that the post autosave doesn't work: the post has no ID and yet its
status is `publish`. That's an impossible combination and the `autosave` action will
return a `NO_AUTOSAVE` error.