Domains: Remove the country code stickiness from the phone input for domains (#24870)
* Domains: Remove the country code stickiness from the phone input for domains

We sometimes get a phone number with mixed up country codes in the domain registration requests. So for example instead of `+44xxx` we get `+144xxx`. We're not 100% sure what the issue is, but one thing that stands out is the "stickiness" of the country if it's choosen from the countries dropdown. So once you choose a country from the dropdown it won't change even if you enter the `+` sign in the phone input. So, I think if we disable the stickiness for the phone input that would probably improve how we handle the phone numbers.

To test - go in the domain contact details form and try entering different country codes in the phone input. Choose different country from the dropdown and then enter different phone with country code in the input field. The country should always match the country code of whatever was entered in the phone input.

Also fixed teh contact-details-form-fields snapshot test.