Post Editor: clean up the QuickSaveButtons component (#24968)
* Post Editor: simplify mapDispatchToProps of Editor History Button

* Post Editor: stop passing `loadRevision` prop to HistoryButton

The only component that needs the `loadRevision` callback is the `EditorRevisionsDialog`
component in the top level `PostEditor`. The `HistoryButton` merely opens a modal dialog,
everything else happens in the modal.

* Post Editor: stop passing Flux post prop to QuickSaveButton

Retrieve the currently edited post from Redux instead.

* Post Editor: remove `showRevisions` prop from QuickSaveButtons

The prop defaults to true and a different value is never passed from parent.