Use geo-tag label as search input value.
In order to simplify the UI, this commit makes the search input serve
double-duty: it is the input for location searches and a display of the
current location's label.

In order to acheive this, we use the Search component's value prop.  In
Search's componentWillReceiveProps() method, however, it takes the value
prop and assigns it to the internal "keyword" property, triggering
another search.  This runs counter to our goal because just want to
display the value rather than doing another search.  The workaround for
this behavior is to detect when the search query matches the value prop
we pass to EditorLocationSearch and just short-circuit the search
process, returning zero results.