Strip out everything but checkbox in EditorLocationOptions.
Now that the search input displays our address label as well, we don't
need the "Description" input in this component.  Because it's now only a
checkbox, we can strip out the popover as well.

We may ultimately be able to get rid of the checkbox and this whole
component as well, depending upon how we decide to handle posts that are
currently "private" because Calypso has so far not been writing the
geo_public meta field at all.