Post Editor: reduxify Post Status optional toggles and buttons
The `EditPostStatus` component displays a few toggles and buttons
(Stick to the front page, Pending review, Revert to draft) that are
displayed conditionally based on the saved and edited post status
(is published? is it page or post? is it private or protected?).

This patch retrieves the saved and edited post info from Redux and stops
using the Flux-based `post` and `savedPost` props.