Privacy/GDPR: add a banner to record sensitive_pixel_option preference (#24820)
This commit adds a banner to record the sensitive_pixel_option preference if the visitor is from a country in the GDPR zone or the visitor's current country cannot be determined. The visitor's choice is stored in a cookie and recorded as a Tracks event. The cookie value is used in isAdTrackingAllowed and when deciding whether to use Google Analytics and Hotjar.