Post Editor: show correct status label for new draft
Fixes the `EditorStatusLabel` component to display a "New Draft" label
instead of "Loading Post..." for new drafts.

If the `post` is `null`, it can mean either that the post is not loaded yet,
or that it's a new draft that hasn't been saved yet.

Use the `isEditorNewPost` selector to distinguish between these cases. That's
more reliable than `post.modified`, too.