Post Editor: reduxify Post Status optional toggles and buttons (#25014)
* Post Editor: reduxify Post Status optional toggles and buttons

The `EditPostStatus` component displays a few toggles and buttons
(Stick to the front page, Pending review, Revert to draft) that are
displayed conditionally based on the saved and edited post status
(is published? is it page or post? is it private or protected?).

This patch retrieves the saved and edited post info from Redux and stops
using the Flux-based `post` and `savedPost` props.

* Fix the getVisibility test

The previous patch has a drive-by fix of return value of `postUtils.getVisibility`,
changing it from `undefined` to `null`: more common value when returning an "unknown"
status. I forgot to update a test for that.