Editor Publish Button: cleanup unit tests
Cleans up the `EditorPublishButton` unit tests that check whether the button click
calls the `onSave` or `onPublish` callback. The tests relied on the `savedPost` prop
which was renamed to `currentPost`. It's funny that the tests still passed although
they were wrong. The reason was that a missing `canPublishPosts` prop has the same
effect as a null `currentPost`: post is saved instead of publish.