Post Editor: cleanup restoring trashed post
Refactors the `onRestore` prop to be a simple handler that's always called the same way,
whether the action is restoring an autosave or restoring a trashed post.

Checking the `delete_posts` capability is not neccessary: to restore a trashed post, the
user needs `edit_posts` and if it was left in the current state, the dialog would be
still displayed, but click on "Restore" would do nothing. Clearly bug or broken code.