Activity Log: Harmonize UI patterns of threats with plugin updates
We have been developing the threat alerts and plugin update
alerts independently and som of the UI has therefore ended
up inconsistent. In this patch I'm attempting to decrease
that gap a little bit so that we don't have alerts that
behave differently.

Notably the cards should no longer be separated from the
overall threat heading and an awkward click behavior on
the header has been fixed.