Signup: Save steps data to persist signup progress
Fixes #24547.

This change makes it so that Signup NavigationLink gets the right previous step name in `getPreviousStepName` for `PlansStep` and `UserStep` components.

Failing to persist the signup step progress here will misdirect the user on clicking the `Back` button. For the `Plans` page, the back button navigates to the `About` step. For the `User` step, the back button navigates to the `Domains` step.