Signup: Save steps data to persist signup progress (#25173)
Fixes #24547. 

Currently, the `Back` buttons in the `Plans` and `User` pages misdirect the user. For the `Plans` page, the back button navigates to the `About` step. For the `User` step, the back button navigates to the `Domains` step.

This change makes it so that the `NavigationLink` in `StepWrapper` gets the right previous step name by persisting the step data after the parent Step components have mounted.